Refactor file operations into separate service
